cancel: Declarando o cancelamento do projeto, amarrei muito meu domain, como é um projeto de estudos irei escrever tudo novamente, usando esse como exemplo

+++ src/domain/value_objects/payment_method/pix.rs
@@ -3,5 +3,5 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct PixData{
-    pub expires_in_seconds: u64,
+    pub expires_in_seconds: u32,
 }
